// frontend/src/types/api.types.ts
export type UserType = {
    _id: string;
    name: string;
    surname: string;
    profilePicture: string | null;
    isActive: boolean;
    lastLogin: string | null;
    role: string;
    navbar: string;
};